+++ src/main/java/pl/krzychuuweb/labelapp/nutritionalvalue/dto/CreateNutritionalValueDTO.java
@@ -7,6 +7,6 @@ public record CreateNutritionalValueDTO(
         @Size(min = 1, max = 80)
         String name,
         @Min(1)
-        Float priority
+        Integer priority
 ) {
 }
